requirements.txt
# Enterprise Detection Engineering Lab
# Python dependencies
# Install: pip install -r requirements.txt

# ── Core ─────────────────────────────────────────────────────────────────────
requests>=2.31.0
pyyaml>=6.0.1

# ── Sigma Rule Conversion ─────────────────────────────────────────────────────
sigma-cli>=0.9.0
pysigma-backend-elasticsearch>=1.1.0
pysigma-backend-splunk>=1.0.0
pysigma-pipeline-windows>=1.2.0
pysigma-pipeline-sysmon>=0.3.0

# ── PDF Report Generation ─────────────────────────────────────────────────────
reportlab>=4.0.0

# ── Elastic Integration ───────────────────────────────────────────────────────
elasticsearch>=8.0.0

# ── Data & Analysis ───────────────────────────────────────────────────────────
python-dateutil>=2.8.2

# ── Development & Testing ─────────────────────────────────────────────────────
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
